[ BUGFIX, REFACTOR ] Fix table nesting | Split actions into two files | Add new config route in server
