+++ client/actions/index.jsx
@@ -4,7 +4,7 @@ export function testRegexInput(challengeId, input, challengeType) {
     regexObject: {
       challengeId,
       input,
-      challengeType: challengeType || 'tutorial',
+      challengeType: challengeType,
     }
   };
 }
@@ -13,30 +13,6 @@ export function inputActionCreator(newInput) {
   return {
     type: 'INPUT-PATTERN-UPDATE',
     newInput,
-  };
-}
-
-export function getAllChallenges() {
-  return (dispatch) => {
-    $.ajax('/regex/challenges').then((data) => {
-      console.log('User Challenge DATAAA', data);
-      dispatch({
-        type: 'GET-CHALLENGES',
-        payload: data,
-      });
-    });
-  };
-}
-
-export function getAllTutorials() {
-  return (dispatch) => {
-    $.ajax('/regex/tutorial').then((data) => {
-      console.log('TUTORIAL DATA', data);
-      dispatch({
-        type: 'GET-TUTORIALS',
-        payload: data,
-      });
-    });
   };
 }
 
@@ -57,21 +33,4 @@ export function postEditTestCase(action, data) {
   }
 }
 
-export function postChallengeActionCreator(postInput) {
-  console.log('Post Input: ', postInput);
-  return (dispatch) => {
-    // To clean postInput of client side properties. 
-    if (postInput.innerMatches) delete postInput.innerMatches;
-    $.ajax({
-      method: 'POST',
-      url: '/regex/challenges',
-      contentType: 'application/json',
-      data: JSON.stringify(postInput),
-    }).then((data) => {
-      dispatch({
-        type: 'POST-CHALLENGE',
-        data,
-      });
-    });
-  };
-}
+
