+++ config.js
@@ -3,3 +3,11 @@ exports = module.exports = {};
 exports.MONGO_URI = 'mongodb://teamUser:regOrDie@ds147975.mlab.com:47975/db_reg';
 
 // $ "mongo ds147975.mlab.com:47975/db_reg -u teamUser -p regOrDie"
+
+exports.GITHUB = {
+	github: {
+	  clientID: `5569d62488f59395fd17`,
+	  clientSecret: `f1f709e882f972a4a50de1d8fe62f25d23d518cf`,
+	  callbackURL: 'http://127.0.0.1:3000/auth/github/callback',
+	}
+}
