+++ server/db/challenges.txt
@@ -1,14 +1,16 @@
 [
 {
-	"order": 0,
-	"name": "User submit",
-	"description": "A user will explain their tite regex challenge",
-	"author": "user123",
-	"difficulty": "superXhard",
+	"name": "Spaces at the beginning and end of strings",
+	"description": "Write a regex that matches whitespace at the beginning and end of strings",
+	"author": "Troy",
+	"difficulty": "2",
 	"testCases": [
-	  { "case": "123", "result": null, "task": "Match", "expectation": true },
-	  { "case": "567", "result": null, "task": "Skip", "expectation": false },
-	  { "case": "hi there", "result": null, "task": "Match", "expectation": true }
+	  { "case": " I have whitespace! ", "result": null, "task": "Match", "expectation": true },
+	  { "case": "I have more whitespace!   ", "result": null, "task": "Match", "expectation": true },
+	  { "case": "   I have some whitespace too!", "result": null, "task": "Match", "expectation": true },
+	  { "case": "Did you say whitespace?", "result": null, "task": "Skip", "expectation": false }
+	  { "case": "Naw", "result": null, "task": "Skip", "expectation": false }
+	  { "case": "Remember me?", "result": null, "task": "Skip", "expectation": false }
 	],
 	"testPassed": false
 },
